Refactor messages into modules and run clippy
