+++ src/lib.rs
@@ -4,7 +4,6 @@
     missing_copy_implementations,
     trivial_casts,
     trivial_numeric_casts,
-    unreachable_pub,
     unsafe_code,
     unstable_features,
     unused_import_braces,
