+++ src/twitch/emotes.rs
@@ -75,7 +75,7 @@ mod tests {
             ("25:0-4", vec![emote!(25, (0..4))]),
             (
                 "1077966:0-6/25:8-12",
-                vec![emote!(1077966, (0..6)), emote!(25, (8..12))],
+                vec![emote!(1_077_966, (0..6)), emote!(25, (8..12))],
             ),
             (
                 "25:0-4,6-10/33:12-19",
